\name{my.first.package-package}
\alias{my.first.package-package}
\alias{my.first.package}
\docType{package}
\title{
\packageTitle{my.first.package}
}
\description{
\packageDescription{my.first.package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{my.first.package}
\packageIndices{my.first.package}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{my.first.package}

Maintainer: \packageMaintainer{my.first.package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
