% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_function.R
\name{square_root}
\alias{square_root}
\title{Title of The First Function}
\usage{
square_root(x = 1)
}
\arguments{
\item{x}{Numeric input parameter. Default = 1.}
}
\value{
The function returns the square root of the input.
}
\description{
First function gets a number and returns the square root of that number.
}
